% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateRespFeatures.R
\name{calculateRespFeatures}
\alias{calculateRespFeatures}
\title{Calculate RRV parameters of the Respiration Signal
Function calculate the time domain parameters of RRV respiratory signal from a
respiration curve.}
\usage{
calculateRespFeatures(
  resp,
  mRR_bool = TRUE,
  stdRR_bool = TRUE,
  RMSSD_bool = TRUE
)
}
\arguments{
\item{resp}{Class Signal object with respiratory signal (output of 'ppgToResp' function)}

\item{mRR_bool}{Specifies if the mRR parameter to be calculated.}

\item{stdRR_bool}{Specifies if the stdRR parameter to be calculated.}

\item{RMSSD_bool}{Specifies if the RMSD parameter to be calculated.}
}
\value{
A S3 class "Signal" object which containing information about a signal
        from param 'resp', with RRV parameters mRR, stdRR and RMSSD
        The "mRR" attribute holds the average length of breathing, while "stdRR"
        standard deviation of the differences between the durations of adjacent breaths
        and RMSSD atribute which contains information about root mean square of
        the differences between the durations of adjacent breaths.

 @examples
          calculateRespFeatures(resp = respObject);
          calculateRespFeatures(resp = respObject, mRR_bool = TRUE, stdRR_bool = FALSE, RMSSD_bool = TRUE);
          calculateRespFeatures(resp = respObject, mRR_bool = FALSE, stdRR_bool = TRUE, RMSSD_bool = TRUE);
          calculateRespFeatures(resp = respObject, mRR_bool = FALSE, stdRR_bool = FALSE, RMSSD_bool = TRUE);
          calculateRespFeatures(resp = respObject, mRR_bool = TRUE, stdRR_bool = FALSE, RMSSD_bool = FALSE);

 @export
}
\description{
Calculate RRV parameters of the Respiration Signal
Function calculate the time domain parameters of RRV respiratory signal from a
respiration curve.
}
