% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardiac_signal.R
\docType{data}
\name{cardiac_signal}
\alias{cardiac_signal}
\title{Synthetic PPG and ECG data}
\format{
A data frame with 2 columns. First column is PPG signal,
second is the ECG signal
}
\usage{
data(cardiac_signal)
}
\description{
The data set contains the synthetic PPG and ECG signals with sampling frequency 500 Hz.
}
\examples{
data(cardiac_signal)
plot(cardiac_signal$V1)
}
\keyword{ECG}
\keyword{PPG}
\keyword{datasets}
