% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPPG.R
\name{loadPPG}
\alias{loadPPG}
\title{Loading the Respiration Signal}
\usage{
loadPPG(file_name, path = "./", first_sample = 0, last_sample = 1000)
}
\arguments{
\item{file_name}{A .csv file with PPG signal}

\item{path}{Signal sampling rate.}

\item{first_sample}{Define the first sample to load}

\item{last_sample}{Define the last sample to load}
}
\value{
A numeric vector contains the PPG signal
}
\description{
Function to load photoplethysmographic (PPG) signal.
}
