% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppgToResp.R
\name{ppgToResp}
\alias{ppgToResp}
\title{Estimating the Respiration Signal}
\usage{
ppgToResp(raw_ppg, sample_rate, normalise_resp = TRUE)
}
\arguments{
\item{raw_ppg}{Vector giving the samples of PPG signal.}

\item{sample_rate}{Signal sampling rate.}

\item{normalise_resp}{Specifies if the normalization to be used.}
}
\value{
A S3 class "Signal" object which containing information about a signal.
        It has four key attributes: "signal", "time", "signal_type", and "graph".
        The "signal" attribute holds the actual signal value, while "time" denotes the time.
        "signal_type" identifies the type of the signal. Lastly, the "graph" attribute contains
        a graphical representation of the signal.

 @examples
          ppgToResp(raw_ppg = ppg_signal, sample_rate = 500);
          ppgToResp(raw_ppg = ppg_signal, sample_rate = 500, normalise_resp = FALSE);
}
\description{
Function estimating the respiratory signal from a photoplethysmographic (PPG)
curve using FM modulation
}
