% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardiac_signal.R
\docType{data}
\name{cardiac_signal}
\alias{cardiac_signal}
\title{Synthetic PPG and ECG data}
\format{
A data frame with 2 columns.
}
\usage{
data(cardiac_signal)
}
\description{
The data set contains the synthetic PPG and ECG signals. First column is PPG signal
second is the ECG signal.Sampling frequency of signal is 500 Hz
}
\examples{
data(cardaic_signal)
plot(cardaic_signal$V1)
}
\keyword{ECG}
\keyword{PPG}
\keyword{datasets}
